requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0　
Django==3.1.2
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-imagekit==4.0.2
idna==2.10
importlib-metadata==2.1.1
oauthlib==3.1.0
pilkit==2.0
pycparser==2.20
Pillow==8.1.0
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
zipp==3.4.0
psycopg2-binary
uwsgi
